feat(US-001): add parallel-batch.ts and rename rectify file
- Implement runParallelBatch() in src/execution/parallel-batch.ts with
  _parallelBatchDeps injection; handles completed stories, failed stories,
  merge conflict rectification, per-story costs (not even-split), and totalCost
- Add src/execution/merge-conflict-rectify.ts as canonical re-export of
  conflict rectification types and rectifyConflictedStory
- Absorb runRectificationPass and ParallelStoryMetrics into parallel-executor.ts
- Delete parallel-executor-rectification-pass.ts (logic absorbed upstream)
- All 21 AC-1..AC-10 tests pass
